src/requirements.txt
https://github.com/pallets/jinja/zipball/2.9-maintenance
pyyaml==3.12
python-json-logger==0.1.8
boto3==1.5.28
pytest==3.3.2
pytest-runner==3.0
validators==0.12.1
python-box==3.1.1
click==6.6
pyinstaller==3.3.1
moto==1.3.1
